+++ CHANGELOG.md
@@ -8,6 +8,9 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [4.72.4]
 ### Changed
 - Automatic test mongod version increased to v7
+### Fixed
+- GnomAD now defaults to hg38 - change build 37 links accordingly
+
 
 
 ## [4.72.3]